Omnia: Add mvneta support for switching PHY
mvneta driver initializes PHY when the NIC is probed and the
PHY is switched up when the NIC is brought up.

Restructure the PHY handling.
Add possibility to define more PHYs to in DT and provide the switch in
sysfs to
/sys/devices/platform/soc/soc:internal-regs/f1034000.ethernet/net/ \
eth{0,1,...}/phy_select file.

The switch accepts the name of DT node that describe which PHY should
be used next time when the NIC is brough up.